Return int instead of VkDraw::Result

+++ src/main.cpp
@@ -4,5 +4,5 @@
 
 int main(const int argc, char** argv) {
 	std::vector<std::string_view> args(argv, argv + argc);
-	return static_cast<int>(VkDraw::run(args));
+	return VkDraw::run(args);
 }
